requirements.txt
flask==3.1.3
chromadb==1.5.5
sentence-transformers==5.3.0
google-genai==1.69.0
python-dotenv==1.2.2
requests==2.32.3
sendgrid==6.11.0
markdown==3.7
